+++ some-api-client/src/main/java/org/spring/boot/metrics/instances/some/api/client/controller/SomeApiDataController.java
@@ -46,11 +46,12 @@ public class SomeApiDataController {
     /**
      * Create some data instance json node.
      *
+     * @param id               the id
      * @param someDataInstance the some data instance
      * @return the json node
      */
     @ApiOperation(
-            value = "Create some data instance",
+            value = "Update or create (if not exists) some data instance",
             notes = "This method allows to create some data instance"
     )
     @ApiResponses(value = {
@@ -59,7 +60,7 @@ public class SomeApiDataController {
                     message = "Successfully completed",
                     examples = @Example(value = @ExampleProperty(
                             mediaType = MediaType.APPLICATION_JSON_VALUE,
-                            value = "{\"id\": 1, \"someString\": \"bla bla\"}"
+                            value = "{\"id\": 1, \"someString\": \"foo\"}"
                     ))
             ),
             @ApiResponse(
@@ -67,8 +68,27 @@ public class SomeApiDataController {
                     message = "Successfully completed",
                     examples = @Example(value = @ExampleProperty(
                             mediaType = MediaType.APPLICATION_JSON_VALUE,
-                            value = "{\"id\": 1, \"someString\": \"bla bla\"}"
+                            value = "{\"id\": 1, \"someString\": \"bar\"}"
                     ))
+            ),
+            @ApiResponse(
+                    code = 400,
+                    message = "Failure uncompleted",
+                    examples = @Example(value = @ExampleProperty(
+                            mediaType = MediaType.APPLICATION_JSON_VALUE,
+                            value = "{" +
+                                    "\"status\": \"BAD_REQUEST\", " +
+                                    "\"timestamp\": \"2021-07-22T09:44:56.972+0000\", " +
+                                    "\"stackTraces\": \"\", " +
+                                    "\"comment\": \"The API execution error - the request data is illegal\", " +
+                                    "\"message\": \"Contact the system administrator\", " +
+                                    "\"details\": \"uri=/some-api/some-data/instances/1;client=0:0:0:0:0:0:0:1\" " +
+                                    "}"
+                    ))
+            ),
+            @ApiResponse(
+                    code = 500,
+                    message = "Service is unreachable"
             )
     })
     @PutMapping(
@@ -80,10 +100,9 @@ public class SomeApiDataController {
             @ApiParam(
                     name="id",
                     value = "The instance of the some data entity",
-                    required = false
+                    required = true
             ) @PathVariable(
-                    name = "id",
-                    required = false
+                    name = "id"
             ) Integer id,
             @ApiParam(
                     name="someDataInstance",
@@ -91,15 +110,8 @@ public class SomeApiDataController {
                     required = true
             ) @RequestBody JsonNode someDataInstance) {
         log.info("Receive request - {} and {}", id, someDataInstance);
-        ResponseEntity<JsonNode> response = null;
-        if (id == null) {
-            response = this.someApiClientService.create(someDataInstance);
-            log.debug("Send create entity response - {}", response);
+        ResponseEntity<JsonNode> response = this.someApiClientService.retrieve(id);
 
-            return response;
-        }
-
-        response = this.someApiClientService.retrieve(id);
         if (response.getStatusCode().equals(HttpStatus.NOT_FOUND)) {
             response = this.someApiClientService.create(someDataInstance);
             log.debug("Send create entity response - {}", response);
@@ -109,6 +121,64 @@ public class SomeApiDataController {
 
         response = this.someApiClientService.update(id, someDataInstance);
         log.debug("Send update entity response - {}", response);
+
+        return response;
+    }
+
+    /**
+     * Delete some data instance response entity.
+     *
+     * @param id the id
+     * @return the response entity
+     */
+    @ApiOperation(
+            value = "Delete some data instance",
+            notes = "This method allows to create some data instance"
+    )
+    @ApiResponses(value = {
+            @ApiResponse(
+                    code = 200,
+                    message = "Successfully completed",
+                    examples = @Example(value = @ExampleProperty(
+                            mediaType = MediaType.APPLICATION_JSON_VALUE,
+                            value = "{\"id\": 1}"
+                    ))
+            ),
+            @ApiResponse(
+                    code = 400,
+                    message = "Failure uncompleted",
+                    examples = @Example(value = @ExampleProperty(
+                            mediaType = MediaType.APPLICATION_JSON_VALUE,
+                            value = "{" +
+                                    "\"status\": \"NOT_FOUND\", " +
+                                    "\"timestamp\": \"2021-07-22T09:44:56.972+0000\", " +
+                                    "\"stackTraces\": \"\", " +
+                                    "\"comment\": \"The API execution error - requested entity is not presented in the service\", " +
+                                    "\"message\": \"There is no SomeData with this id - 111\", " +
+                                    "\"details\": \"uri=/some-api/some-data/instances/1;client=0:0:0:0:0:0:0:1\" " +
+                                    "}"
+                    ))
+            ),
+            @ApiResponse(
+                    code = 500,
+                    message = "Service is unreachable"
+            )
+    })
+    @DeleteMapping(
+            path = "/{id}",
+            produces = "application/json"
+    )
+    public ResponseEntity<JsonNode> deleteSomeDataInstance(
+            @ApiParam(
+                    name="id",
+                    value = "The instance of the some data entity",
+                    required = true
+            ) @PathVariable(
+                    name = "id"
+            ) Integer id) {
+        log.info("Receive request - {}", id);
+        ResponseEntity<JsonNode> response = this.someApiClientService.delete(id);
+        log.debug("Send delete entity response - {}", response);
 
         return response;
     }