+++ some-api-executing/some-api-crud-executing.sh
@@ -8,5 +8,7 @@ do
     curl -X PUT "http://some-api-client:8082/some-api/some-data/instances/$iteration" -H "accept: application/json" -H "Content-Type: application/json" -d "{\"id\": $iteration, \"someString\": \"foo\"}"
     sleep 5
     curl -X PUT "http://some-api-client:8082/some-api/some-data/instances/$iteration" -H "accept: application/json" -H "Content-Type: application/json" -d "{\"id\": $iteration, \"someString\": \"bar\"}"
-    sleep 25
+    sleep 5
+    curl -X DELETE "http://some-api-client:8082/some-api/some-data/instances/$iteration" -H "accept: application/json"
+    sleep 20
 done